Added Field Value Calculation
The Field value function as been implemented. The function now returns the total value of the field. Some small tweaks have been done to other classes, this was mainly changing unnecessary doubles to integers.

+++ src/students/items/Apples.java
@@ -4,7 +4,7 @@ public class Apples extends Food{
 	// represented as 'a' or 'A'
 	
 	private static int appleCount = 0;
-	public double purchacePrice = 2;
+	public int purchacePrice = 2;
 	
 	public Apples() {
 		super(3, 5, 3);
