+++ src/students/items/Food.java
@@ -2,7 +2,7 @@ package students.items;
 
 public abstract class Food extends Item{
 	
-	public Food(int maturationAge, int deathAge, double value) {
+	public Food(int maturationAge, int deathAge, int value) {
 		super(maturationAge,deathAge,value);
 	}
