+++ src/students/items/Grain.java
@@ -4,7 +4,7 @@ public class Grain extends Food{
 	// represented as 'g' or 'G'
 	
 	private static int grainCount = 0;
-	public double purchacePrice = 1;// No clue if this is even supposed to be a thing (no direct mention?)
+	public int purchacePrice = 1;// No clue if this is even supposed to be a thing (no direct mention?)
 
 	public Grain() {
 		// Mature: 2, Death: 6, Value: 2
